+++ src/error.ts
@@ -1,0 +1,13 @@
+import createError from '@fastify/error'
+
+export class SignatureValidationError extends createError(
+  'LINE_SIGNATURE_VALIDATION_FAILED',
+  'Signature validation failed',
+) {
+  constructor(
+    public message: string,
+    public signature?: string,
+  ) {
+    super()
+  }
+}
